Move App to a dedicated submodule
Cleanup only, no functional changes.
